Extract MCP tool definitions into SPMCPToolDefinitions
The tool catalogue was locked inside a private extension on SPMCPServer,
which the Unit Tests target cannot compile: SPMCPServer reads ObjC
constants from SPConstants.h and that target has no bridging header.

Moves the definitions verbatim into a dependency-free namespace built
into both targets, the same shape as SPMCPHTTP and SPMCPFavorite, so the
advertised schemas can be asserted in unit tests.
